+++ data.h
@@ -1,7 +1,7 @@
-#ifndef REDUNDANTDATA_H
-#define REDUNDANTDATA_H 1
+#ifndef QR_DATA_H
+#define QR_DATA_H 1
 
-#include "packedbits.h"
+#include "bits.h"
 #include "shared.h"
 
 /* Creates blocks of data/ecc from string,
@@ -10,7 +10,7 @@
 typedef struct qrdata_s qrdata_t;
 
 __attribute__((__nonnull__)) int
-create_qrdata(qrdata_t** self, const pbits_t* __restrict__ bits,
+create_qrdata(qrdata_t** self, const bits_t* __restrict__ bits,
               const qrinfo_t* const __restrict__ info);
 __attribute__((__nonnull__)) void
 delete_qrdata(qrdata_t** self);